siae_evaluations: Move administrative criteria selection to siae_select_criteria
Previously, the form was responsible for selecting the administrative
criteria. With the certification, additional data must be fetched, and
the form does not have enough information to efficiently fetch the
additional data. Moving the data collection to the view makes it easier
to prefetch and use the prefetched data immediately.
