+++ tests/www/eligibility_views/test_forms.py
@@ -1,19 +1,10 @@
-import pytest
-from dateutil.relativedelta import relativedelta
-from django.utils import timezone
-
 from itou.companies.enums import CompanyKind
 from itou.eligibility.enums import (
-    ADMINISTRATIVE_CRITERIA_LEVEL_2_REQUIRED_FOR_SIAE_KIND,
     AdministrativeCriteriaKind,
-    AdministrativeCriteriaLevel,
 )
-from itou.eligibility.models import AdministrativeCriteria, EligibilityDiagnosis
+from itou.eligibility.models import AdministrativeCriteria
 from itou.www.eligibility_views.forms import AdministrativeCriteriaForm
-from itou.www.siae_evaluations_views.forms import AdministrativeCriteriaEvaluationForm
 from tests.companies.factories import CompanyFactory
-from tests.job_applications.factories import JobApplicationFactory
-from tests.users.factories import JobSeekerFactory
 
 
 class TestAdministrativeCriteriaForm:
@@ -172,55 +163,3 @@ class TestAdministrativeCriteriaForm:
 
         assert not AdministrativeCriteriaForm(is_authorized_prescriber=False, siae=company, data={}).is_valid()
         assert AdministrativeCriteriaForm(is_authorized_prescriber=False, siae=None, data={}).is_valid()
-
-
-class TestAdministrativeCriteriaEvaluationForm:
-    def test_job_application(self):
-        company = CompanyFactory(with_membership=True)
-        user = company.members.first()
-
-        job_seeker = JobSeekerFactory()
-
-        eligibility_diagnosis = EligibilityDiagnosis.create_diagnosis(
-            job_seeker,
-            author=user,
-            author_organization=company,
-            administrative_criteria=[
-                AdministrativeCriteria.objects.filter(level=AdministrativeCriteriaLevel.LEVEL_1).first()
-            ]
-            + [AdministrativeCriteria.objects.filter(level=AdministrativeCriteriaLevel.LEVEL_2).first()],
-        )
-
-        job_application = JobApplicationFactory(
-            with_approval=True,
-            to_company=company,
-            sender_company=company,
-            eligibility_diagnosis=eligibility_diagnosis,
-            hiring_start_at=timezone.localdate() - relativedelta(months=2),
-        )
-
-        form = AdministrativeCriteriaEvaluationForm(user, company, job_application=job_application)
-
-        assert 2 == len(form.fields)
-        assert (
-            AdministrativeCriteria.objects.filter(level=AdministrativeCriteriaLevel.LEVEL_1).first().key
-            in form.fields.keys()
-        )
-        assert (
-            AdministrativeCriteria.objects.filter(level=AdministrativeCriteriaLevel.LEVEL_2).first().key
-            in form.fields.keys()
-        )
-
-    @pytest.mark.parametrize("kind", CompanyKind.siae_kinds())
-    def test_num_level2_admin_criteria(self, kind):
-        company = CompanyFactory(kind=kind, with_membership=True)
-        user = company.members.first()
-
-        job_application = JobApplicationFactory(
-            with_approval=True,
-            to_company=company,
-            sender_company=company,
-            hiring_start_at=timezone.localdate() - relativedelta(months=2),
-        )
-        form = AdministrativeCriteriaEvaluationForm(user, company, job_application=job_application)
-        assert ADMINISTRATIVE_CRITERIA_LEVEL_2_REQUIRED_FOR_SIAE_KIND[kind] == form.num_level2_admin_criteria
